Added the possibility to send the userip keyword on requests.
This is part of the Google API http://code.google.com/apis/ajaxsearch/documentation/reference.html#_restUrlBase and is defined as "This argument supplies the IP address of the end-user on whose behalf the request is being made. Requests that include it are less likely to be mistaken for abuse. In choosing to utilize this parameter, please be sure that you're in compliance with any local laws, including any laws relating to disclosure of personal information being sent."